custom exception has written

+++ superbook2/db/db_user.py
@@ -1,3 +1,5 @@
+from fastapi import HTTPException, status
+
 from superbook2.db.hash import Hash
 from sqlalchemy.orm.session import Session
 from schema import UserBase
@@ -21,11 +23,16 @@ def get_all_users(db: Session):
 
 
 def get_user(db: Session, id: int):
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'user id {id}')
+    return 'ok'
 
 
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'user id {id}')
     user.update({
         DbUser.username: request.username,
         DbUser.email: request.email,
@@ -37,6 +44,8 @@ def update_user(db: Session, id: int, request: UserBase):
 
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'user id {id}')
     db.delete(user)
     db.commit()
     return 'ok'
